:recycle: (tarot catalog): move files, modulur struc, fine tune degign
