+++ src/app/_components/_tarot/CardCatalogViewer/TCImageTab.tsx
@@ -1,0 +1,23 @@
+import Image from "next/image";
+import { Card, CardBody } from "@nextui-org/card";
+export default function TCImageDisplay({
+  imgUrl,
+  title,
+}: {
+  imgUrl: string;
+  title: string;
+}) {
+  return (
+    <Card>
+      <CardBody>
+        <Image
+          className="mx-auto"
+          src={imgUrl}
+          alt={title}
+          width={350}
+          height={150}
+        />
+      </CardBody>
+    </Card>
+  );
+};
