+++ src/app/_components/_tarot/CardCatalogViewer/index.tsx
@@ -1,0 +1,3 @@
+import CardCatalogContainer from "./CardCatalogContainer";
+
+export default CardCatalogContainer;
